requirements.txt
asgiref==3.5.0
Babel==2.9.1
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-money==2.1.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
importlib-metadata==4.11.3
Markdown==3.3.6
phonenumbers==8.12.46
py-moneyed==1.2
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
zipp==3.7.0
